+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
+
+
+#1.1.0
+Pebble Media Secretary: Now has additional commands for remotely restarting sonarr/radarr.
+services-manager-server: a node.js script for launching shell scripts for remotely resetting sonarr/radarr